src/requirements.txt
langgraph
requests
python-dotenv
google-genai
pandas
streamlit
feedparser